Re-wrote analysis script
For some reasons, we don’t save all the data at the end of the 5x5
block - instead data are saved per condition. This will change - but
for now this script sorts everything based on the TF and SF listed in
the finalData. Note - this makes it impossible to perform proper stats…